+++ common/fuzzing/proto_to_carbon.h
@@ -15,8 +15,10 @@ namespace Carbon {
 // buffer is invalid (like a variable declaration with an empty `name` field).
 // This is done to reduce the number of inputs the fuzzer framework generates
 // when trying to produce lexically valid source.
-auto ProtoToCarbon(const Fuzzing::Carbon& proto) -> std::string;
+auto ProtoToCarbon(const Fuzzing::Carbon& proto, bool maybe_add_main)
+    -> std::string;
 
+// Parses the textproto into a proto object.
 auto ParseCarbonTextProto(const std::string& contents)
     -> ErrorOr<Fuzzing::Carbon>;
