+++ common/fuzzing/proto_to_carbon_test.cpp
@@ -2,21 +2,18 @@
 // Exceptions. See /LICENSE for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "explorer/fuzzing/fuzzer_util.h"
+#include "common/fuzzing/proto_to_carbon.h"
 
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
 
-#include <fstream>
-
-#include "common/fuzzing/proto_to_carbon.h"
-#include "llvm/ADT/StringRef.h"
-#include "llvm/Support/raw_ostream.h"
+#include "common/error.h"
+#include "common/fuzzing/carbon.pb.h"
 
 namespace Carbon::Testing {
 namespace {
 
-TEST(FuzzerUtilTest, ParseAndExecute) {
+TEST(FuzzerUtilTest, ProtoToCarbon) {
   const ErrorOr<Fuzzing::Carbon> carbon_proto = ParseCarbonTextProto(R"(
     compilation_unit {
       package_statement { package_name: "P" }
@@ -42,16 +39,34 @@ TEST(FuzzerUtilTest, ParseAndExecute) {
       }
     })");
   ASSERT_TRUE(carbon_proto.ok());
-  const ErrorOr<int> result = ParseAndExecute(*carbon_proto);
-  ASSERT_TRUE(result.ok()) << "Execution failed: " << result.error();
-  EXPECT_EQ(*result, 0);
+  static constexpr char SourceCode[] = R"(// Generated by proto_to_carbon.
+package P api;
+
+fn Main() -> i32
+{
+return 0;
 }
 
-TEST(FuzzerUtilTest, GetRunfilesFile) {
-  EXPECT_THAT(*GetRunfilesFile("carbon/explorer/data/prelude.carbon"),
-              testing::EndsWith("/prelude.carbon"));
-  EXPECT_THAT(GetRunfilesFile("nonexistent-file").error().message(),
-              testing::EndsWith("doesn't exist"));
+)";
+  EXPECT_THAT(ProtoToCarbon(*carbon_proto, /*maybe_add_main=*/false),
+              testing::Eq(SourceCode));
+  EXPECT_THAT(ProtoToCarbon(*carbon_proto, /*maybe_add_main=*/true),
+              testing::Eq(SourceCode));
+}
+
+TEST(FuzzerUtilTest, ParseCarbonTextProtoWithUnknownField) {
+  const ErrorOr<Fuzzing::Carbon> carbon_proto = ParseCarbonTextProto(R"(
+    compilation_unit {
+      garbage: "value"
+      declarations {
+        choice {
+          name {
+            name: "Ch"
+          }
+        }
+      }
+    })");
+  ASSERT_FALSE(carbon_proto.ok());
 }
 
 }  // namespace
