+++ explorer/fuzzing/BUILD
@@ -16,21 +16,18 @@ cc_library(
     ],
 )
 
-cc_library(
-    name = "fuzzer_util",
+cc_binary(
+    name = "ast_to_proto",
     testonly = 1,
-    srcs = ["fuzzer_util.cpp"],
-    hdrs = ["fuzzer_util.h"],
+    srcs = ["ast_to_proto_main.cpp"],
     deps = [
-        "//common:check",
+        ":ast_to_proto_lib",
+        "//common:bazel_working_dir",
         "//common:error",
         "//common/fuzzing:carbon_cc_proto",
-        "//common/fuzzing:proto_to_carbon_lib",
-        "//explorer/interpreter:exec_program",
-        "//explorer/interpreter:trace_stream",
+        "//explorer/ast",
+        "//explorer/common:arena",
         "//explorer/syntax",
-        "//explorer/syntax:prelude",
-        "@bazel_tools//tools/cpp/runfiles",
         "@com_google_protobuf//:protobuf_headers",
         "@llvm-project//llvm:Support",
     ],
@@ -50,34 +47,32 @@ cc_test(
     deps = [
         ":ast_to_proto_lib",
         "//common/fuzzing:carbon_cc_proto",
+        "//common/fuzzing:proto_to_carbon_lib",
         "//explorer/syntax",
         "@com_google_googletest//:gtest",
         "@com_google_protobuf//:protobuf_headers",
-        "@llvm-project//llvm:Support",
     ],
 )
 
-cc_binary(
-    name = "fuzzverter",
+cc_library(
+    name = "fuzzer_util",
     testonly = 1,
-    srcs = ["fuzzverter.cpp"],
+    srcs = ["fuzzer_util.cpp"],
+    hdrs = ["fuzzer_util.h"],
     deps = [
-        ":ast_to_proto_lib",
-        ":fuzzer_util",
+        "//common:check",
         "//common:error",
         "//common/fuzzing:carbon_cc_proto",
         "//common/fuzzing:proto_to_carbon_lib",
-        "//explorer/common:error_builders",
-        "//explorer/common:nonnull",
+        "//explorer/ast",
+        "//explorer/interpreter:exec_program",
+        "//explorer/interpreter:trace_stream",
         "//explorer/syntax",
+        "//explorer/syntax:prelude",
+        "@bazel_tools//tools/cpp/runfiles",
         "@com_google_protobuf//:protobuf_headers",
         "@llvm-project//llvm:Support",
     ],
-)
-
-filegroup(
-    name = "fuzzer_corpus_files",
-    srcs = glob(["fuzzer_corpus/**"]),
 )
 
 cc_test(
@@ -90,50 +85,6 @@ cc_test(
         ":fuzzer_util",
         "//common:gtest_main",
         "//common/fuzzing:proto_to_carbon_lib",
-        "@com_google_googletest//:gtest",
-        "@com_google_protobuf//:protobuf_headers",
-        "@llvm-project//llvm:Support",
-    ],
-)
-
-cc_test(
-    name = "proto_to_carbon_test",
-    srcs = ["proto_to_carbon_test.cpp"],
-    args = [
-        "$(locations //explorer:standard_libraries)",
-        "$(locations //explorer/testdata:carbon_files)",
-    ],
-    data = [
-        "//explorer:standard_libraries",
-        "//explorer/testdata:carbon_files",
-    ],
-    deps = [
-        ":ast_to_proto_lib",
-        "//common/fuzzing:carbon_cc_proto",
-        "//common/fuzzing:proto_to_carbon_lib",
-        "//explorer/syntax",
-        "@com_google_googletest//:gtest",
-        "@com_google_protobuf//:protobuf_headers",
-        "@llvm-project//llvm:Support",
-    ],
-)
-
-cc_test(
-    name = "clone_test",
-    srcs = ["clone_test.cpp"],
-    args = [
-        "$(locations //explorer:standard_libraries)",
-        "$(locations //explorer/testdata:carbon_files)",
-    ],
-    data = [
-        "//explorer:standard_libraries",
-        "//explorer/testdata:carbon_files",
-    ],
-    deps = [
-        ":ast_to_proto_lib",
-        "//common/fuzzing:carbon_cc_proto",
-        "//explorer/ast",
-        "//explorer/syntax",
         "@com_google_googletest//:gtest",
         "@com_google_protobuf//:protobuf_headers",
         "@llvm-project//llvm:Support",
