+++ explorer/fuzzing/fuzzer_util.cpp
@@ -19,8 +19,7 @@
 
 namespace Carbon {
 
-auto Internal::GetRunfilesFile(const std::string& file)
-    -> ErrorOr<std::string> {
+auto GetRunfilesFile(const std::string& file) -> ErrorOr<std::string> {
   using bazel::tools::cpp::runfiles::Runfiles;
   std::string error;
   // `Runfiles::Create()` fails if passed an empty `argv0`.
@@ -37,14 +36,14 @@ auto Internal::GetRunfilesFile(const std::string& file)
 }
 
 auto ParseAndExecute(const Fuzzing::Carbon& carbon) -> ErrorOr<int> {
-  const std::string source = ProtoToCarbon(carbon);
+  const std::string source = ProtoToCarbon(carbon, /*maybe_add_main=*/true);
 
   Arena arena;
   CARBON_ASSIGN_OR_RETURN(AST ast,
                           ParseFromString(&arena, "Fuzzer.carbon", source,
                                           /*parser_debug=*/false));
   const ErrorOr<std::string> prelude_path =
-      Internal::GetRunfilesFile("carbon/explorer/data/prelude.carbon");
+      GetRunfilesFile("carbon/explorer/data/prelude.carbon");
   // Can't do anything without a prelude, so it's a fatal error.
   CARBON_CHECK(prelude_path.ok()) << prelude_path.error();
